requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-filter>=23.0
PyMySQL>=1.1.0
cryptography>=41.0.0
Pillow>=10.0.0
